Solution to Prime Number Checker

+++ Prime Number Checker@@67a4d240c8f64283936a1643/code.c
@@ -1,4 +1,4 @@
-float isPrime(int num){
+int isPrime(int num){
     if (num<=1){
         return 0;
     }
